Play sounds when boxes collide
The sounds are far too long and needs to be redone. But it's a start.
There are also some parameters for tweaking when the sounds play.
Right now it's a mess.

Also fixes a Physics bug where moving around would cause weird
errors to happen - essentially we were relying on the old physics
location when calculating new physics locations, rather than
correctly loading them from the Hit Result.